requirements.txt
langchain
langgraph
langchain-google-genai
google-generativeai
pydantic
requests
beautifulsoup4
pandas
python-dotenv